requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2022.1
cx-Oracle==8.3.0
Django==4.2.5
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap-form==3.4
django-crispy-forms==2.0
django-filter==23.2
geographiclib==2.0
geopy==2.4.0
mysqlclient==2.2.0
pydantic==2.3.0
pymemcache==4.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
djangorestframework~=3.14.0
django-debug-toolbar~=4.2.0
python-dotenv~=1.0.0
